Added the whole project
This project was developed as part of the final exam for the 13th module in the "Applied Programmer" qualification program.  The system features both front-end and back-end components, supporting user-friendly interaction and efficient database handling. It was created over the course of a 3-year specialized programming course aimed at high school students.